Introduce store types, set from command line
